voice/backend/requirements.txt
# Core ML/AI
torch>=2.0.0
transformers>=4.53.0
accelerate>=0.20.0
torch-audio>=2.0.0
torchvision>=0.15.0

# Gemma 3n specific
sentencepiece>=0.1.99
protobuf>=3.20.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0

# Audio Processing
speechrecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.11
librosa>=0.10.0

# Web Interface
streamlit>=1.28.0
gradio>=3.50.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Mobile Development
kivy>=2.2.0
buildozer>=1.5.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.4.0
loguru>=0.7.0
tqdm>=4.66.0
requests>=2.31.0

# Development
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0
pre-commit>=3.5.0
